pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexusrag"
version = "0.1.0"
description = "An open-source framework for building autonomous AI agents that reason over complex, multimodal data."
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "streamlit>=1.28.0",
    "unstructured>=0.10.10",
    "sentence-transformers>=2.2.2",
    "chromadb>=0.4.15",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "pymupdf>=1.23.5",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "python-docx>=0.8.11",
    "beautifulsoup4>=4.12.2",
    "openai>=1.3.5",
    "cohere>=4.37",
    "pinecone-client>=2.2.4",
    "weaviate-client>=3.24.2",
    "anthropic>=0.18.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "google-generativeai>=0.3.0",
    "ollama>=0.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
dev = [
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/your-username/NexusRAG"
Repository = "https://github.com/your-username/NexusRAG"
Documentation = "https://github.com/your-username/NexusRAG/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["nexusrag*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
